[Websocket] - refactor - przeniesienie springow'ych klas konfiguracyjnych do wspolnego pakietu configuration
