+++ src/main/java/com/github/britter/springbootherokudemo/configuration/Application.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.britter.springbootherokudemo;
+package com.github.britter.springbootherokudemo.configuration;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.*;
 
 @SpringBootApplication
 public class Application {
