+++ src/main/java/com/github/britter/springbootherokudemo/configuration/ApplicationInitializer.java
@@ -1,4 +1,4 @@
-package com.github.britter.springbootherokudemo;
+package com.github.britter.springbootherokudemo.configuration;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
