+++ src/main/java/com/github/britter/springbootherokudemo/configuration/i18n/I18nConfiguration.java
@@ -1,0 +1,19 @@
+package com.github.britter.springbootherokudemo.configuration.i18n;
+
+import org.springframework.context.*;
+import org.springframework.context.annotation.*;
+import org.springframework.context.support.*;
+import org.springframework.web.servlet.config.annotation.*;
+
+@Configuration
+public class I18nConfiguration extends WebMvcConfigurerAdapter {
+
+    @Bean
+    public MessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasename("i18n/messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+
+}
